# mo-marketplace-api/mo-marketplace-openapi.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User registered, returns JWT"},"409":{"description":"Email already registered"}},"summary":"Register a new user","tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Returns JWT access_token"},"401":{"description":"Invalid credentials"}},"summary":"Login and get JWT token","tags":["auth"]}},"/products":{"post":{"operationId":"ProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"Product created"}},"security":[{"bearer":[]}],"summary":"Create a new product","tags":["products"]},"get":{"operationId":"ProductsController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Get all products (paginated)","tags":["products"]}},"/products/{id}":{"get":{"operationId":"ProductsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Product found"},"404":{"description":"Product not found"}},"summary":"Get a product by ID","tags":["products"]},"patch":{"operationId":"ProductsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update a product","tags":["products"]},"delete":{"operationId":"ProductsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Product deleted"}},"security":[{"bearer":[]}],"summary":"Delete a product","tags":["products"]}},"/variants":{"post":{"operationId":"VariantsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVariantDto"}}}},"responses":{"201":{"description":"Variant created with combination_key"},"409":{"description":"Duplicate variant combination"}},"security":[{"bearer":[]}],"summary":"Create variant for a product","tags":["variants"]}},"/variants/product/{productId}":{"get":{"operationId":"VariantsController_findByProduct","parameters":[{"name":"productId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get all variants for a product","tags":["variants"]}},"/variants/{id}/stock":{"patch":{"operationId":"VariantsController_updateStock","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update stock quantity for a variant","tags":["variants"]}},"/variants/{id}":{"delete":{"operationId":"VariantsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete a variant","tags":["variants"]}}},"info":{"title":"MO Marketplace API","description":"Full-Stack Assessment — Products, Variants, JWT Auth","version":"1.0","contact":{}},"tags":[{"name":"auth","description":""},{"name":"products","description":""},{"name":"variants","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"password123","minLength":6}},"required":["email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"password123"}},"required":["email","password"]},"CreateProductDto":{"type":"object","properties":{"name":{"type":"string","example":"Classic T-Shirt","description":"Product name"},"description":{"type":"string","example":"A comfortable everyday t-shirt"},"basePrice":{"type":"number","example":29.99,"description":"Base price in USD"},"category":{"type":"string","example":"Clothing"},"imageUrl":{"type":"string","example":"https://example.com/image.jpg"}},"required":["name","basePrice"]},"UpdateProductDto":{"type":"object","properties":{"name":{"type":"string","example":"Classic T-Shirt","description":"Product name"},"description":{"type":"string","example":"A comfortable everyday t-shirt"},"basePrice":{"type":"number","example":29.99,"description":"Base price in USD"},"category":{"type":"string","example":"Clothing"},"imageUrl":{"type":"string","example":"https://example.com/image.jpg"}}},"CreateVariantDto":{"type":"object","properties":{"productId":{"type":"string","example":"uuid-of-product"},"color":{"type":"string","example":"red"},"size":{"type":"string","example":"M"},"material":{"type":"string","example":"cotton"},"priceModifier":{"type":"number","example":5,"description":"Extra price on top of base"},"stockQuantity":{"type":"number","example":100,"description":"Stock quantity (0 = out of stock)"}},"required":["productId","color","size","material","stockQuantity"]}}}}